new: entity.collisions.fluid: use chunk.block_counting to skip fluid drag calculations when there is no fluid nearby

+++ src/main/java/me/jellysquid/mods/lithium/mixin/chunk/block_counting/ChunkSectionMixin.java
@@ -34,7 +34,7 @@ public abstract class ChunkSectionMixin implements SectionFlagHolder {
 
     @Override
     public boolean getFlag(IndexedBlockStatePredicate indexedBlockStatePredicate) {
-        return this.countsByFlag[indexedBlockStatePredicate.getIndex()] != 0;
+        return this.countsByFlag[indexedBlockStatePredicate.getIndex()] != (short) 0;
     }
 
     @Redirect(
